SDM: removeVolumes/removeImage commands and garbage collection
Add support for garbage collecting half-baked volumes and use the
garbage collector to implement the removeVolumes and removeImage
commands.

To remove a volume we simply atomically append a suffix to the target
volume's name that will be recognized by the garbage collector and then
directly invoke the garbage collector.

removeVolumes is a synchronous command and failure or success is
indicated by the returned status code.  If the command returns
successfully, then it can be assumed that all requested volumes were
deleted.  If an error is returned, then some of the volumes may have
been deleted.  The engine may resubmit the command with the same volume
list and missing volumes will be logged and skipped.

removeImage is similar to removeVolumes except a volume list is not
specified.  The command will attempt to remove all volumes in the image.
On succes, the image is removed, on failure, some volumes may have been
removed.

Garbage collection occurs in two modes: in-line and batch

Inline Garbage Collection:
When volumes are removed from an image they are atomically renamed so
the garbage collector will recognize them.  In order to immediately
correct the image structure and reclaim storage space, the inline
garbage collector (garbageCollectVolumes) is invoked for the image.

Batch Garbage Collection:
In cases where operations fail before the in-line GC runs and in other
cases (such as removing the last volume in an image), the storage will
accumulate garbage which should be periodically cleaned up.  This patch
creates the Batch GC (garbageCollectStorageDomain) but does not arrange
for it to be called yet.  It should probably be run whenever a storage
domain is activated and periodically thereafter.

Change-Id: I66f846171bcc9920f05d38dc37d2f30a0443e209
Signed-off-by: Adam Litke <alitke@redhat.com>