sdm: introduce the new createVolumeContainer command
This command is used to create a new container to hold a volume.  This
is a metadata operation and requires the SDM lock on the target storage
domain.  The command works to achieve the same goals as the old
createVolume command except for preallocation which will be a separate
API.  Since we are no longer using the asyncTask framework or rollbacks,
we must maintain the ability to discover incomplete and failed attempts
so that they can be garbage collected later.

This operation is synchronous.  Success or failure is indicated by the
returned status code.

Garbage collection notes:
While under construction a suffix is added to the volume name.  This
identifies the volume as a candidate for garbage collection.  Once all
setup is complete the volume is atomically renamed to its actual name.

Change-Id: If3ea624f81a04a830cdcbcfe2f777cdd7bd26c08
Signed-off-by: Adam Litke <alitke@redhat.com>