ImageManifest: introduce ImageManifest class
In order to safely share storage code between the traditional SPM verbs
and the new SDM verbs we must refactor the code from the Image class.
We want to share the parts which deal with image "metadata" (ie. the
repoPath and imageDir, volumeChain, etc).  These elements will be moved
into a new ImageManifest class while the SPM operations will remain in
the Image class.

As a first step, create the new class and add repoPath and getImageDir.
Many of the other candidate functions rely on the VolumeMetadata
refactoring and will be added later.

Change-Id: Ic7a78aef10902b9baf6f88d2e053e2232e2a1267
Signed-off-by: Adam Litke <alitke@redhat.com>