StorageDomainManifest: Move clusterLock into Manifest
The clusterLock will be used by SDM verbs (as a per-domain lock) so it
needs to be moved into the Manifest so that it can be shared.

Moving the lock requires us to instantiate it in the
StorageDomainManifest constructor which requires some reordering of the
initialization steps.  For testing, all BlockStorageDomainManifest tests
must create the manifest with a mocked LVM.  In addition, all manifest
tests must ensure that the metadata contains a 'VERSION' key.
volumeTests.py was removed and the metadataSlot tests were moved into a
new class inside manifest_tests.py.

Change-Id: I2c019a7fed6ba793d2b3a2459af221b1abbff2a3
Signed-off-by: Adam Litke <alitke@redhat.com>