storage: Factor apparentsize adjustment into a reusable function
When creating a new volume sometimes the actual size is larger than the
requested size due to underlying storage limitations (such as physical
extent granularity in LVM).  In these cases, the new volume's size must
be adjusted to match the actual size.  The SDM volume creation code
would also like to use this function so extract it into the
VolumeMetadata class.

Change-Id: I7395850e127b941fc50ae8d5659868a2499b5076
Signed-off-by: Adam Litke <alitke@redhat.com>