storage: Refactor new volume metadata generation
XXX: Actually, _putMetadata is simple enough we can probably apply the
gcSuffix in a sister method in the FileVolumeArtifacts class.

To prepare for garbage-collectible volume metadata needed by the SDM
flows, we need to break out the functionality of creating new volume
metadata.  Introduce makeMetadata() which builds a metadata dict from a
given set of args.  Also, modify _putMetadata to allow an optional
extension to be supplied.

Change-Id: I028996b3a5967f96a103f0b03d6b8f82b25da888
Signed-off-by: Adam Litke <alitke@redhat.com>