doc: Add SDM overview document
In an effort to provide better documentation for the storage code this
document outlines the major design points of SDM storage management.

The storage/docs directory and its content is inspired by the
Documentation directory in the Linux kernel source code.  It provides
information that isn't appropriate for placement directly in the source
code.  As such, we must be careful to ensure that changes to the code
that affect the documentation result in updates to these documents.

Change-Id: I6e7b070a2587189d05268053068f1746c025aab2
Signed-off-by: Adam Litke <alitke@redhat.com>