sdm: Add create_volume job
This is the first SDM verb to be wired up using all of the
infrastructure (jobs and getHostJobs API, VolumeArtifacts, etc.).  HSM
creates a create_volume.Job and schedules it using the existing
TaskManager thread pool.  Eventually we can switch to using the executor
framework but that can be for a later patch.

Change-Id: Ia614059f52c9625da7841ea9fbca2b2f2375cd75
Signed-off-by: Adam Litke <alitke@redhat.com>