storage: Factor out new volume size adjustment
The volume create flow contains logic used to check and inflate (if
necessary) the size of a newly created volume if the underlying device
is larger.  This logic should be encapsulated into a helper function
which can be tested.  It is also reusable by the SDM VolumeArtifacts so
let's put it into the VolumeManifest.

Change-Id: I70a785071c89515b0e9f5f4eba2ae0ffa11f80ae
Signed-off-by: Adam Litke <alitke@redhat.com>