tests: Test storagetestlib FakeEnv functions
The storagetestlib module provides functionality to create fake file and
block storage environments which help to unit test storage code.  As
this code has gotten more complex we need to test that it is actually
working correctly itself.

The new storagetestlibTests module tests both file and block
environments.  We check that the temporary directory housing the fake
environment has the correct layout and that metadata for the domain and
volumes 'persists' to the storage.

Change-Id: I65ab8c6419dfa1181efff2a533c00589ba5b8263
Signed-off-by: Adam Litke <alitke@redhat.com>